# internal/api/genclient/genclient.gen.go
// Package genclient provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package genclient

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/gorilla/mux"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Account defines model for Account.
type Account struct {
	AvailableMinor     *int64              `json:"available_minor,omitempty"`
	BalanceMinor       *int64              `json:"balance_minor,omitempty"`
	Currency           *string             `json:"currency,omitempty"`
	Iban               *string             `json:"iban,omitempty"`
	Id                 *openapi_types.UUID `json:"id,omitempty"`
	IsDefault          *bool               `json:"is_default,omitempty"`
	Kind               *string             `json:"kind,omitempty"`
	Status             *string             `json:"status,omitempty"`
	TransferLimitMinor *int64              `json:"transfer_limit_minor,omitempty"`
	UserId             *openapi_types.UUID `json:"user_id,omitempty"`
}

// CreateTransferRequest defines model for CreateTransferRequest.
type CreateTransferRequest struct {
	AmountMinor   int64              `json:"amount_minor"`
	CreditAccount openapi_types.UUID `json:"credit_account"`
	DebitAccount  openapi_types.UUID `json:"debit_account"`
	Description   *string            `json:"description,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Error   string  `json:"error"`
	Message *string `json:"message,omitempty"`
}

// Health defines model for Health.
type Health struct {
	Status  *string `json:"status,omitempty"`
	Version *string `json:"version,omitempty"`
}

// LedgerEntry defines model for LedgerEntry.
type LedgerEntry struct {
	AccountIban       *string             `json:"account_iban,omitempty"`
	AccountId         *openapi_types.UUID `json:"account_id,omitempty"`
	AmountMinor       *int64              `json:"amount_minor,omitempty"`
	BalanceAfter      *int64              `json:"balance_after,omitempty"`
	CounterpartyIban  *string             `json:"counterparty_iban,omitempty"`
	CounterpartyOwner *string             `json:"counterparty_owner,omitempty"`
	Currency          *string             `json:"currency,omitempty"`
	Description       *string             `json:"description,omitempty"`
	Direction         *string             `json:"direction,omitempty"`
	Id                *openapi_types.UUID `json:"id,omitempty"`
	PostedAt          *time.Time          `json:"posted_at,omitempty"`
	SignedAmount      *int64              `json:"signed_amount,omitempty"`
	TransferId        *openapi_types.UUID `json:"transfer_id,omitempty"`
	TransferKind      *string             `json:"transfer_kind,omitempty"`
	TransferStatus    *string             `json:"transfer_status,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	ExpiresAt *time.Time `json:"expires_at,omitempty"`

	// Token JWT access token (HS256). Send as 'Authorization: Bearer <token>'.
	Token     *string             `json:"token,omitempty"`
	TokenType *string             `json:"token_type,omitempty"`
	UserId    *openapi_types.UUID `json:"user_id,omitempty"`
}

// ReasonRequest defines model for ReasonRequest.
type ReasonRequest struct {
	Reason *string `json:"reason,omitempty"`
}

// StatusResponse defines model for StatusResponse.
type StatusResponse struct {
	Id     *openapi_types.UUID `json:"id,omitempty"`
	Status *string             `json:"status,omitempty"`
}

// Transfer defines model for Transfer.
type Transfer struct {
	AmountMinor     *int64              `json:"amount_minor,omitempty"`
	CreditAccountId *openapi_types.UUID `json:"credit_account_id,omitempty"`
	Currency        *string             `json:"currency,omitempty"`
	DebitAccountId  *openapi_types.UUID `json:"debit_account_id,omitempty"`
	Description     *string             `json:"description,omitempty"`
	FailureReason   *string             `json:"failure_reason,omitempty"`
	Id              *openapi_types.UUID `json:"id,omitempty"`
	Kind            *string             `json:"kind,omitempty"`
	PostedAt        *time.Time          `json:"posted_at,omitempty"`
	RequestedAt     *time.Time          `json:"requested_at,omitempty"`
	ReversesId      *openapi_types.UUID `json:"reverses_id,omitempty"`
	Status          *string             `json:"status,omitempty"`
}

// TransferResult defines model for TransferResult.
type TransferResult struct {
	Status     *string             `json:"status,omitempty"`
	TransferId *openapi_types.UUID `json:"transfer_id,omitempty"`
	WasReplay  *bool               `json:"was_replay,omitempty"`
}

// Cursor defines model for Cursor.
type Cursor = time.Time

// Id defines model for Id.
type Id = openapi_types.UUID

// IdempotencyKey defines model for IdempotencyKey.
type IdempotencyKey = string

// Limit defines model for Limit.
type Limit = int

// bearerAuthContextKey is the context key for bearerAuth security scheme
type bearerAuthContextKey string

// sessionCookieContextKey is the context key for sessionCookie security scheme
type sessionCookieContextKey string

// GetAccountLedgerParams defines parameters for GetAccountLedger.
type GetAccountLedgerParams struct {
	Cursor *Cursor `form:"cursor,omitempty" json:"cursor,omitempty"`
	Limit  *Limit  `form:"limit,omitempty" json:"limit,omitempty"`
}

// CreateTransferParams defines parameters for CreateTransfer.
type CreateTransferParams struct {
	IdempotencyKey IdempotencyKey `json:"Idempotency-Key"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// CreateTransferJSONRequestBody defines body for CreateTransfer for application/json ContentType.
type CreateTransferJSONRequestBody = CreateTransferRequest

// CancelTransferJSONRequestBody defines body for CancelTransfer for application/json ContentType.
type CancelTransferJSONRequestBody = ReasonRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get an account with ledger + available balance
	// (GET /accounts/{id})
	GetAccount(w http.ResponseWriter, r *http.Request, id Id)
	// Account statement (cursor-paginated, running balance)
	// (GET /accounts/{id}/ledger)
	GetAccountLedger(w http.ResponseWriter, r *http.Request, id Id, params GetAccountLedgerParams)
	// Verify credentials (bcrypt). Establishes a session in the portal.
	// (POST /auth/login)
	Login(w http.ResponseWriter, r *http.Request)
	// Liveness/version probe
	// (GET /health)
	Health(w http.ResponseWriter, r *http.Request)
	// Create a transfer (auto-posts by default). Idempotent.
	// (POST /transfers)
	CreateTransfer(w http.ResponseWriter, r *http.Request, params CreateTransferParams)
	// Get a transfer
	// (GET /transfers/{id})
	GetTransfer(w http.ResponseWriter, r *http.Request, id Id)
	// Cancel a pending transfer (releases the hold). Idempotent.
	// (POST /transfers/{id}/cancel)
	CancelTransfer(w http.ResponseWriter, r *http.Request, id Id)
	// Post a pending transfer (deferred settlement). Idempotent.
	// (POST /transfers/{id}/post)
	PostTransfer(w http.ResponseWriter, r *http.Request, id Id)
	// List a user's accounts
	// (GET /users/{id}/accounts)
	ListUserAccounts(w http.ResponseWriter, r *http.Request, id Id)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAccount operation middleware
func (siw *ServerInterfaceWrapper) GetAccount(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAccount(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetAccountLedger operation middleware
func (siw *ServerInterfaceWrapper) GetAccountLedger(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAccountLedgerParams

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "cursor", r.URL.Query(), &params.Cursor, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "cursor"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cursor", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAccountLedger(w, r, id, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Health operation middleware
func (siw *ServerInterfaceWrapper) Health(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Health(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTransfer operation middleware
func (siw *ServerInterfaceWrapper) CreateTransfer(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateTransferParams

	headers := r.Header

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey IdempotencyKey
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err = fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTransfer(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTransfer operation middleware
func (siw *ServerInterfaceWrapper) GetTransfer(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTransfer(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CancelTransfer operation middleware
func (siw *ServerInterfaceWrapper) CancelTransfer(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CancelTransfer(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTransfer operation middleware
func (siw *ServerInterfaceWrapper) PostTransfer(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTransfer(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListUserAccounts operation middleware
func (siw *ServerInterfaceWrapper) ListUserAccounts(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", mux.Vars(r)["id"], &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListUserAccounts(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{})
}

type GorillaServerOptions struct {
	BaseURL          string
	BaseRouter       *mux.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r *mux.Router) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r *mux.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options GorillaServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = mux.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.HandleFunc(options.BaseURL+"/accounts/{id}", wrapper.GetAccount).Methods(http.MethodGet)

	r.HandleFunc(options.BaseURL+"/accounts/{id}/ledger", wrapper.GetAccountLedger).Methods(http.MethodGet)

	r.HandleFunc(options.BaseURL+"/auth/login", wrapper.Login).Methods(http.MethodPost)

	r.HandleFunc(options.BaseURL+"/health", wrapper.Health).Methods(http.MethodGet)

	r.HandleFunc(options.BaseURL+"/transfers", wrapper.CreateTransfer).Methods(http.MethodPost)

	r.HandleFunc(options.BaseURL+"/transfers/{id}", wrapper.GetTransfer).Methods(http.MethodGet)

	r.HandleFunc(options.BaseURL+"/transfers/{id}/cancel", wrapper.CancelTransfer).Methods(http.MethodPost)

	r.HandleFunc(options.BaseURL+"/transfers/{id}/post", wrapper.PostTransfer).Methods(http.MethodPost)

	r.HandleFunc(options.BaseURL+"/users/{id}/accounts", wrapper.ListUserAccounts).Methods(http.MethodGet)

	return r
}